+++ example/append.js
@@ -1,9 +1,0 @@
-var append = require('..').append;
-var path = require('path');
-
-console.log(
-    ['x.css', 'y.css'].map(
-        append(1, ['.css'], path.basename)
-    )
-);
-