+++ example/prepend.js
@@ -1,9 +1,0 @@
-var prepend = require('..').prepend;
-var path = require('path');
-
-console.log(
-    ['x.css', 'y.css'].map(
-        prepend(1, ['/path/to/css'], path.resolve)
-    )
-);
-